qai_hub_models_bk/requirements.txt
Pillow>10,<12
numpydoc==1.9.0
gdown==4.7.1
gitpython==3.1.42
huggingface_hub>=0.34.0,<1.0 # 0.34 contains a fix for long windows paths
ipython==8.12.3
numpy<3
onnx>=1.17,<1.19 # aimet-onnx==2.20.0 requires onnx>=1.17
onnxruntime>=1.19,<1.23 # 1.19 is allowed for compatibility with AIMET-ONNX
ai-edge-litert>=2.0.2,<2.1.2; python_version<'3.13' and sys_platform != 'win32' # We use LiteRT instead of TF Lite because the equivalent TF Lite package is no longer maintained and does not have the TF Lite updates beyond 2023.
opencv-python>4,<5
pandas>2,<2.3
prettytable==3.11.0
requests_toolbelt==1.0.0
schema==0.7.5
torch>=2.4,<2.12.0
tabulate==0.9.0
torchvision>=0.19,<0.27.0
typing-extensions>=4.12.2
tqdm>=4.66
pyarrow==21.0.0
qai_hub>=0.47
datasets==4.1.1
ruamel-yaml==0.18.10
filelock>=3.16.1
pydantic>=2,<3
pydantic_yaml==1.4.0
scipy>=1.8.1,<2 # required by all ImageNet models to load the ImageNet dataset. This is a dependency of torchvision that is only included in an extra. TODO: Figure out how we can remove this, since not all models need it.
packaging>24.2
coverage>=7.6.1
platformdirs>=4.5.0,<5.0.0
boto3>=1.34,<1.36
botocore>=1.34,<1.36
boto3-stubs[s3]==1.36.9
